+++ CarND-Extended-Kalman-Filter-Project/src/kalman_filter.cpp
@@ -35,16 +35,16 @@ void KalmanFilter::Update(const VectorXd &z) {
 
     //new estimate
     x_ = x_ + (K * y);
-    long x_size = x_.size();
+    long long x_size = x_.size();
     MatrixXd I = MatrixXd::Identity(x_size, x_size);
     P_ = (I - K * H_) * P_;
 
 }
 
 void KalmanFilter::UpdateEKF(const VectorXd &z) {
-    float rho = sqrt(x_[0]*x_[0] + x_[1]*x_[1]);
-    float phi = atan2(x_[1], x_[0]);
-    float rho_dot;
+    double rho = sqrt(x_[0]*x_[0] + x_[1]*x_[1]);
+    double phi = atan2(x_[1], x_[0]);
+    double rho_dot;
 
     if(rho != 0.0){
         rho_dot = (x_[0]*x_[2] + x_[1]*x_[3])/rho;
@@ -63,7 +63,7 @@ void KalmanFilter::UpdateEKF(const VectorXd &z) {
 
     //new estimate
     x_ = x_ + (K * y);
-    long x_size = x_.size();
+    long long x_size = x_.size();
     MatrixXd I = MatrixXd::Identity(x_size, x_size);
     P_ = (I - K * H_) * P_;
